# openapi.yaml
openapi: 3.0.3
info:
  title: Heating equipment selection service
  version: "1.0.0"
  license:
    name: ISC
    url: https://opensource.org/license/isc-license-txt/
servers:
  - url: /

# Публичные эндпоинты (calc, catalog, presets) — без auth. /api/v1/projects — JWT (см. ProjectsBearerAuth).
security: []
paths:
...
  /api/v1/admin/feedback:
    get:
      summary: List client feedback (admin only)
      operationId: listAdminFeedback
      security:
        - ProjectsBearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          description: Opaque cursor returned by the previous page.
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/FeedbackStatus"
        - name: type
          in: query
          schema:
            type: string
            enum: [bug, contact]
      responses:
        "200":
          description: Feedback page ordered by newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFeedbackListResponse"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: PROJECTS_AUTH_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: ADMIN_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: MongoDB unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/admin/feedback/stream:
    get:
      summary: Stream feedback events to an open admin dashboard
      operationId: streamAdminFeedback
      description: |
        Authenticated Server-Sent Events stream. The server sends heartbeat comments and
        `feedback.created` events. Clients must reload the REST list after reconnecting;
        MongoDB remains the source of truth.
      security:
        - ProjectsBearerAuth: []
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          description: PROJECTS_AUTH_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: ADMIN_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: MongoDB unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/admin/feedback/{id}:
    patch:
      summary: Change client feedback status (admin only)
      operationId: patchAdminFeedback
      security:
        - ProjectsBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: feedback._id (ObjectId hex)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminFeedbackPatchBody"
      responses:
        "200":
          description: Updated feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFeedbackPatchResponse"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: PROJECTS_AUTH_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: ADMIN_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: FEEDBACK_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: MongoDB unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/admin/users/{id}:
    patch:
      summary: Update user role or subscription (admin only)
      operationId: patchAdminUser
      security:
        - ProjectsBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: users._id (ObjectId hex)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserPatchBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserPatchResponse"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: PROJECTS_AUTH_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: ADMIN_REQUIRED / INVALID_USER_ROLE / INVALID_SUBSCRIPTION_TIER
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: USER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: MongoDB недоступна
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  securitySchemes:
    ProjectsBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Clerk OIDC access JWT (Bearer). Backend: JOSE + JWKS (RS256/ES256), без Clerk SDK.
        Frontend: Clerk SDK → getToken() → Authorization header.
        Identity: JWT.sub → users.providerUserId → users._id → projects.ownerId.

        Runtime env (Clerk): AUTH_JWKS_URI, AUTH_ISSUER, AUTH_AUDIENCE, AUTH_PROVIDER=clerk.
        AUTH_JWKS_URI и AUTH_JWT_SECRET взаимоисключающие — для runtime только JWKS.

        Production и PROJECTS_AUTH_ENABLED=true: JWT обязателен; сервер не стартует без JWKS-конфига.

        Dev без PROJECTS_AUTH_ENABLED: JWT опционален.

        Коды ошибок — см. ProjectsAuthErrorCode.
  schemas:
...
    AdminUserPatchBody:
      $ref: ./components/schemas/AdminUserPatchBody.yaml
    AdminUserPatchResponse:
      $ref: ./components/schemas/AdminUserPatchResponse.yaml
    FeedbackStatus:
      $ref: ./components/schemas/FeedbackStatus.yaml
...
    AdminFeedbackListResponse:
      $ref: ./components/schemas/AdminFeedbackListResponse.yaml
    AdminFeedbackPatchBody:
      $ref: ./components/schemas/AdminFeedbackPatchBody.yaml
    AdminFeedbackPatchResponse:
      $ref: ./components/schemas/AdminFeedbackPatchResponse.yaml
...
    ErrorEnvelope:
      $ref: ./components/schemas/ErrorEnvelope.yaml
...